requirements.txt
boto3==1.26.12
celery==5.2.7
Django==4.1.3
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
django-storages==1.13.1
ipython==8.6.0
requests==2.28.1
pika==1.3.1
pillow==9.3.0
psycopg2--binary==2.9.5
python-dotenv==0.21.0